requirements.txt
requests
pandas
psycopg2-binary
sqlalchemy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client